+++ public/app/listeners/index.js
@@ -2,10 +2,14 @@ const loadSpace = require('./loadSpace');
 const saveSpace = require('./saveSpace');
 const getSpace = require('./getSpace');
 const getSpaces = require('./getSpaces');
+const showSyncSpacePrompt = require('./showSyncSpacePrompt');
+const syncSpace = require('./syncSpace');
 
 module.exports = {
   loadSpace,
   saveSpace,
   getSpace,
   getSpaces,
+  showSyncSpacePrompt,
+  syncSpace,
 };